implement rest request from metal archives

+++ src/main/groovy/com/metalr2/butler/config/resttemplate/HttpClientConfigConstants.groovy
@@ -8,7 +8,7 @@ interface HttpClientConfigConstants {
   int MAX_LOCALHOST_CONNECTIONS = 80
 
   // Keep alive
-  int DEFAULT_KEEP_ALIVE_TIME = 20 * 1000 // 20 sec
+  long DEFAULT_KEEP_ALIVE_TIME = 20 * 1000 // 20 sec
 
   // Timeouts
   int CONNECTION_TIMEOUT = 30 * 1000 // 30 sec, the time for waiting until a connection is established
@@ -16,6 +16,6 @@ interface HttpClientConfigConstants {
   int SOCKET_TIMEOUT     = 60 * 1000 // 60 sec, the time for waiting for data
 
   // Idle connection monitor
-  int IDLE_CONNECTION_WAIT_TIME = 30 * 1000 // 30 sec
+  long IDLE_CONNECTION_WAIT_TIME = 30 * 1000 // 30 sec
 
 }
